% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_icd10-data.R
\docType{data}
\name{dd_icd10}
\alias{dd_icd10}
\title{ICD-10 Mortality Data Dictionary}
\format{A \code{data.frame} object containing the column names and descriptions
for the ICD-10 code table "mort"
\describe{
  \item{code}{code used in the mortality table}
  \item{detailed}{detailed ICD-10 code list}
  \item{cause}{general cause of death}
}}
\source{
\href{https://www.who.int/healthinfo/statistics/mortality_rawdata/en/}{WHO Mortality Database}
}
\usage{
data(dd_icd10)
}
\description{
Data dictionary for the ICD-10 codes used in the mortality table, extracted from Table 8 in the
documentation Word document "Documentation_1May2019.doc"
}
\examples{
data(dd_icd10)
head(dd_icd10)
}
\keyword{datasets}
