% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_pop-data.R
\docType{data}
\name{dd_pop}
\alias{dd_pop}
\title{WHO Population Data Dictionary}
\format{A \code{data.frame} object containing the column names and descriptions
for the population data table "pop"
\describe{
  \item{column_name}{the column name used for the WHO population table}
  \item{content}{the description of the column name}
  \item{type}{expected data type for the column}
}}
\source{
\href{https://www.who.int/healthinfo/statistics/mortality_rawdata/en/}{WHO Mortality Database}
}
\usage{
data(dd_pop)
}
\description{
Data dictionary for the population raw files extracted from the 
documentation Word document "Documentation_1May2019.doc"
}
\examples{
data(dd_pop)
head(dd_pop)
}
\keyword{datasets}
