% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{download_who}
\alias{download_who}
\title{Download WHO Mortality Data}
\usage{
download_who(dest, data = c("docs", "mort", "pop", "ccode", "all"))
}
\arguments{
\item{dest}{a \code{character} file path of the destintation directory to download the .zip files}

\item{data}{to specify which files to download, takes values of "docs", "mort", "pop", "ccode", "all". Default value downloads the documentation files}
}
\value{
the same value as the \code{\link[utils]{download.file}}
}
\description{
This function downloads the raw WHO mortality data based on the links on the webpage \url{https://www.who.int/healthinfo/statistics/mortality_rawdata/en/}
}
\details{
The \code{data} parameter allows the user to download a specific set of data files.
The \code{docs} downloads the documentation files associated with the raw WHO mortality data, this includes the "documentation.zip", "notes.zip", and "availability.zip" files
The \code{mort} downloads the two ICD-10 mortality .zip files, "Morticd10_part1.zip" and "Morticd10_part2.zip"
The \code{pop} downloads the WHO population data, "pop.zip"
The \code{ccode} downloads the country code table "country_codes.zip"
The \code{all} downloads all the abovementioned .zip files at once
}
\examples{
\dontrun{
download_who(dest = getwd(), data = "ccode")
}
}
