% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{extract_who}
\alias{extract_who}
\title{Extract WHO Mortality Data}
\usage{
extract_who(path, dest)
}
\arguments{
\item{path}{a \code{character} file path where the downloaded WHO .zip files exist}

\item{dest}{a \code{character} destination file path to extract, default value will be \code{path}}
}
\value{
a named \code{logical} vector of the files that were successfully extracted
}
\description{
Extract WHO Mortality Data
}
\examples{
\dontrun{
download_who(dest = getwd(), data = "ccode")
download_who(dest = getwd(), data = "docs")
extract_who(path = getwd(), dest = getwd(), convert = TRUE)
}
}
