% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dd.R
\name{get_dd}
\alias{get_dd}
\title{Import Data Dictionary for WHO Mortality and Population Tables}
\usage{
get_dd(type = c("mort", "pop"), ...)
}
\arguments{
\item{type}{a single \code{character} specifying to import either the mortality or population data dictionary}

\item{...}{other parameters pass to \code{\link{read_csv}}.}
}
\value{
a \code{data.frame} of the data dictionary with mappings of field names and descriptions
}
\description{
This function imports the data dictionaries for the mortality and population tables. These were extracted from the documentation word document "Documentation_1May2019.doc".
}
\examples{
\dontrun{
get_dd(type = "mort")
get_dd(type = "pop")
}
}
