% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{import_who}
\alias{import_who}
\title{Import WHO Mortality Data}
\usage{
import_who(path, data = c("mort", "pop", "ccode", "notes"))
}
\arguments{
\item{path}{a \code{character} file path where the extracted .csv WHO files exist}

\item{data}{a \code{character} to specify which files to import into R, must be explicitly defined}
}
\value{
a \code{tibble} object of the specified data
}
\description{
This function imports the csv WHO mortality files, extracted from downloaded zip files from a local directory
}
\details{
The \code{data} parameter allows the user to download a specific set of data files.
The \code{mort} imports the two ICD-10 mortality csv files, "Morticd10_part1.zip" and "Morticd10_part2.zip" and combines them into a single \code{tibble} object
The \code{pop} imports the WHO population csv file
The \code{ccode} imports the country code csv file
The \code{notes} imports the notes csv file
}
\examples{
\dontrun{
download_who(dest = getwd(), data = "ccode")
extract_who(path = getwd(), dest = getwd(), convert = TRUE)
ccode_df <- import_who(path = getwd(), data = "ccode")
head(ccode_df)
}
}
