% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{proc_who_mort}
\alias{proc_who_mort}
\title{Process Raw WHO Mortality Data}
\usage{
proc_who_mort(x)
}
\arguments{
\item{x}{a \code{data.frame} of the original csv WHO mortality table, this is the output of \code{\link{import_who}}}
}
\value{
a \code{data.frame} object with modified column data types
}
\description{
This function processes the original WHO mortality csv file to correct data types for each column.
}
\examples{
\dontrun{
mort <- import_who(path = getwd(), data = "mort")
mort <- proc_who_mort(mort)
}
}
