% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{snakecase_cols}
\alias{snakecase_cols}
\title{Convert Column Names to Lower Snake Case}
\usage{
snakecase_cols(x)
}
\arguments{
\item{x}{a \code{data.frame} of the original csv WHO mortality table, this is the output of \code{\link{import_who}}}
}
\value{
a \code{data.frame} with snake case column names
}
\description{
This function converts all column names of input \code{data.frame} to lower snake case.
}
\examples{
\dontrun{
ccodes <- import_who(path = getwd(), data = "ccode")
names(ccodes)
ccodes <- snakecase_cols(ccodes)
names(ccodes)
}
}
